+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+### Removed
+- `encode.standard` removed, users should use the encodings classes directly.
+- `core.topphatt_standard` removed, use `encode.TernaryTree.topphatt` and `encode.TernaryTree.encode_topphatt` instead.
+- `TernaryTree.from_hamiltonian_coefficients`, users can construct a `FermionHamiltonain` and use `TernaryTree(fham.n_modes)`.
+
 ## [0.7.3]
 ### Added
 - Water 6-31G benchmark data for performance testing.